remove old gfa-style implementations of node depth, clean up test directory and Makefile, add new test cases
