+++ .gitignore
@@ -1,7 +1,12 @@
 __pycache__/
 *.py[cod]
 
-k.gfa
-DRB1-3123.gfa
-*.og
+test/*.og
+test/*.gfa
+test/*.out
+
+test/basic/*.og
+test/basic/*.out
+
 *.json
+og_to_gfa.py
