+++ test/basic/ex1.gfa
@@ -1,0 +1,6 @@
+H	VN:Z:1.0
+S	1	A
+L	1	+	2	+	0M
+S	2	C
+L	2	+	2	+	0M
+P	path1	1+,2+,2+	*
