+++ test/basic/ex2.gfa
@@ -1,0 +1,14 @@
+H	VN:Z:1.0
+S	1	AA
+L	1	+	2	+	0M
+S	2	AC
+L	2	+	3	+	0M
+L	2	+	4	+	0M
+S	3	AG
+L	3	+	2	+	0M
+L	3	+	5	+	0M
+S	4	AT
+L	4	+	5	+	0M
+S	5	CG
+P	path0	1+,2+,3+,2+,4+,5+	*
+P	path1	1+,2+,3+,5+	*
