--- test/k.out
@@ -1,16 +1,0 @@
-#node.id	depth	depth.uniq
-1	2	2
-2	1	1
-3	1	1
-4	1	1
-5	1	1
-6	2	2
-7	1	1
-8	1	1
-9	2	2
-10	1	1
-11	1	1
-12	2	2
-13	1	1
-14	1	1
-15	2	2
