+++ test/turnt.toml
@@ -1,11 +1,11 @@
 [envs.baseline]
 command = "odgi build -g {filename} -o temp.og; odgi depth -i temp.og -d; rm temp.og"
 
-[envs.py1]
-command = "python3 ../depth1.py {filename}"
+# [envs.py1]
+# command = "python3 ../depth1.py {filename}"
 
-[envs.py-og]
-command = "odgi build -g {filename} -o temp.og; python3 ../depth3.py temp.og; rm temp.og"
+[envs.py-odgi]
+command = "odgi build -g {filename} -o temp.og; python3 ../depth.py temp.og; rm temp.og"
 
 # [envs.dahlia]
 # command = "fud exec ../depth.fuse --from dahlia --to dat --through verilog -s verilog.data ../depth.fuse.data | python3 ../process.py json"
